[action] Renamed the invdyn actions (default name is for condensed formulation)
