+++ bindings/python/crocoddyl/multibody/actions/free-invdyn.cpp
@@ -1,7 +1,7 @@
 ///////////////////////////////////////////////////////////////////////////////
 // BSD 3-Clause License
 //
-// Copyright (C) 2021, University of Edinburgh
+// Copyright (C) 2021-2022, University of Edinburgh, Heriot-Watt University
 // Copyright note valid unless otherwise stated in individual files.
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
@@ -17,24 +17,21 @@ void exposeDifferentialActionFreeInvDynamics() {
   bp::scope().attr("yes") = 1;
   bp::scope().attr("no") = 0;
   {
-    bp::register_ptr_to_python<boost::shared_ptr<DifferentialActionModelFreeInvDynamics> >();
+    bp::register_ptr_to_python<boost::shared_ptr<DifferentialActionModelFreeInvDynamics>>();
     bp::scope model_outer =
-        bp::class_<DifferentialActionModelFreeInvDynamics, bp::bases<DifferentialActionModelAbstract> >(
+        bp::class_<DifferentialActionModelFreeInvDynamics, bp::bases<DifferentialActionModelAbstract>>(
             "DifferentialActionModelFreeInvDynamics",
             "Differential action model for free inverse dynamics in multibody systems.\n\n"
-            "This class implements  the dynamics using Recursive Newton Euler Algorithm (RNEA) as an equality "
+            "This class implements the dynamics using Recursive Newton Euler Algorithm (RNEA) as an equality "
             "constraint.\n"
             "The stack of cost and constraint functions are implemented in\n"
             "ConstraintModelManager() and CostModelSum(), respectively.",
             bp::init<boost::shared_ptr<StateMultibody>, boost::shared_ptr<ActuationModelAbstract>,
-                     boost::shared_ptr<CostModelSum>, bp::optional<boost::shared_ptr<ConstraintModelManager> > >(
+                     boost::shared_ptr<CostModelSum>, bp::optional<boost::shared_ptr<ConstraintModelManager>>>(
                 bp::args("self", "state", "actuation", "costs", "constraints"),
                 "Initialize the free inverse-dynamics action model.\n\n"
-                "It describes the kinematic evolution of the multibody system without any contact,\n"
-                "and imposes an inverse-dynamics (equality) constraint. Additionally, it computes\n"
-                "the cost and extra constraint values associated to this state and control pair.\n"
-                "Note that the name `rnea` in the ConstraintModelManager is reserved to store\n"
-                "the inverse-dynamics constraint\n."
+                "It describes the kinematic evolution of the multibody system and computes the\n"
+                "needed torques using inverse dynamics.\n"
                 ":param state: multibody state\n"
                 ":param actuation: abstract actuation model\n"
                 ":param costs: stack of cost functions\n"
@@ -81,54 +78,55 @@ void exposeDifferentialActionFreeInvDynamics() {
                                             bp::return_value_policy<bp::return_by_value>()),
                           "entire constraint model");
 
-    bp::register_ptr_to_python<boost::shared_ptr<DifferentialActionModelFreeInvDynamics::ResidualModelRnea> >();
+    bp::register_ptr_to_python<boost::shared_ptr<DifferentialActionModelFreeInvDynamics::ResidualModelActuation>>();
 
-    bp::class_<DifferentialActionModelFreeInvDynamics::ResidualModelRnea, bp::bases<ResidualModelAbstract> >(
-        "ResidualModelRnea",
-        "This residual function is defined as r = tau - RNEA, where\n"
-        "tau is extracted from the control vector and RNEA evaluates the joint torque using\n"
-        "q, q_dot, acc values.",
+    bp::class_<DifferentialActionModelFreeInvDynamics::ResidualModelActuation, bp::bases<ResidualModelAbstract>>(
+        "ResidualModelActuation",
+        "This residual function enforces the torques of under-actuated joints (e.g., floating-base\n"
+        "joints) to be zero. We compute these torques and their derivatives using RNEA inside \n"
+        "DifferentialActionModelFreeInvDynamics.",
         bp::init<boost::shared_ptr<StateMultibody>, std::size_t>(bp::args("self", "state", "nu"),
-                                                                 "Initialize the RNEA residual model.\n\n"
-                                                                 ":param nu: dimension of control vector"))
-        .def<void (DifferentialActionModelFreeInvDynamics::ResidualModelRnea::*)(
+                                                                 "Initialize the actuation residual model.\n\n"
+                                                                 ":param state: state description\n"
+                                                                 ":param nu: dimension of the joint torques"))
+        .def<void (DifferentialActionModelFreeInvDynamics::ResidualModelActuation::*)(
             const boost::shared_ptr<ResidualDataAbstract>&, const Eigen::Ref<const Eigen::VectorXd>&,
             const Eigen::Ref<const Eigen::VectorXd>&)>(
-            "calc", &DifferentialActionModelFreeInvDynamics::ResidualModelRnea::calc,
+            "calc", &DifferentialActionModelFreeInvDynamics::ResidualModelActuation::calc,
             bp::args("self", "data", "x", "u"),
-            "Compute the RNEA residual.\n\n"
+            "Compute the actuation residual.\n\n"
             ":param data: residual data\n"
             ":param x: state point (dim. state.nx)\n"
             ":param u: control input (dim. nu)")
-        .def<void (DifferentialActionModelFreeInvDynamics::ResidualModelRnea::*)(
+        .def<void (DifferentialActionModelFreeInvDynamics::ResidualModelActuation::*)(
             const boost::shared_ptr<ResidualDataAbstract>&, const Eigen::Ref<const Eigen::VectorXd>&)>(
             "calc", &ResidualModelAbstract::calc, bp::args("self", "data", "x"))
-        .def<void (DifferentialActionModelFreeInvDynamics::ResidualModelRnea::*)(
+        .def<void (DifferentialActionModelFreeInvDynamics::ResidualModelActuation::*)(
             const boost::shared_ptr<ResidualDataAbstract>&, const Eigen::Ref<const Eigen::VectorXd>&,
             const Eigen::Ref<const Eigen::VectorXd>&)>(
-            "calcDiff", &DifferentialActionModelFreeInvDynamics::ResidualModelRnea::calcDiff,
+            "calcDiff", &DifferentialActionModelFreeInvDynamics::ResidualModelActuation::calcDiff,
             bp::args("self", "data", "x", "u"),
-            "Compute the Jacobians of the RNEA residual.\n\n"
+            "Compute the Jacobians of the actuation residual.\n\n"
             "It assumes that calc has been run first.\n"
             ":param data: action data\n"
             ":param x: state point (dim. state.nx)\n"
             ":param u: control input (dim. nu)\n")
-        .def<void (DifferentialActionModelFreeInvDynamics::ResidualModelRnea::*)(
+        .def<void (DifferentialActionModelFreeInvDynamics::ResidualModelActuation::*)(
             const boost::shared_ptr<ResidualDataAbstract>&, const Eigen::Ref<const Eigen::VectorXd>&)>(
             "calcDiff", &ResidualModelAbstract::calcDiff, bp::args("self", "data", "x"))
-        .def("createData", &DifferentialActionModelFreeInvDynamics::ResidualModelRnea::createData,
+        .def("createData", &DifferentialActionModelFreeInvDynamics::ResidualModelActuation::createData,
              bp::with_custodian_and_ward_postcall<0, 2>(), bp::args("self", "data"),
-             "Create the RNEA residual data.\n\n"
+             "Create the actuation residual data.\n\n"
              "Each residual model has its own data that needs to be allocated. This function\n"
-             "returns the allocated data for the RNEA residual.\n"
+             "returns the allocated data for the actuation residual.\n"
              ":param data: shared data\n"
              ":return residual data.");
   }
 
-  bp::register_ptr_to_python<boost::shared_ptr<DifferentialActionDataFreeInvDynamics> >();
+  bp::register_ptr_to_python<boost::shared_ptr<DifferentialActionDataFreeInvDynamics>>();
 
   bp::scope data_outer =
-      bp::class_<DifferentialActionDataFreeInvDynamics, bp::bases<DifferentialActionDataAbstract> >(
+      bp::class_<DifferentialActionDataFreeInvDynamics, bp::bases<DifferentialActionDataAbstract>>(
           "DifferentialActionDataFreeInvDynamics", "Action data for the free inverse-dynamics system.",
           bp::init<DifferentialActionModelFreeInvDynamics*>(bp::args("self", "model"),
                                                             "Create free inverse-dynamics action data.\n\n"
@@ -150,15 +148,15 @@ void exposeDifferentialActionFreeInvDynamics() {
                                         bp::return_value_policy<bp::return_by_value>()),
                         "constraint data");
 
-  bp::register_ptr_to_python<boost::shared_ptr<DifferentialActionDataFreeInvDynamics::ResidualDataRnea> >();
+  bp::register_ptr_to_python<boost::shared_ptr<DifferentialActionDataFreeInvDynamics::ResidualDataActuation>>();
 
-  bp::class_<DifferentialActionDataFreeInvDynamics::ResidualDataRnea, bp::bases<ResidualDataAbstract> >(
-      "ResidualDataRnea", "Data for RNEA residual.\n\n",
-      bp::init<DifferentialActionModelFreeInvDynamics::ResidualModelRnea*, DataCollectorAbstract*>(
+  bp::class_<DifferentialActionDataFreeInvDynamics::ResidualDataActuation, bp::bases<ResidualDataAbstract>>(
+      "ResidualDataActuation", "Data for actuation residual.\n\n",
+      bp::init<DifferentialActionModelFreeInvDynamics::ResidualModelActuation*, DataCollectorAbstract*>(
           bp::args("self", "model", "data"),
-          "Create RNEA residual data.\n\n"
-          ":param model: RNEA residual model\n"
-          ":param data: shared data")[bp::with_custodian_and_ward<1, 2, bp::with_custodian_and_ward<1, 3> >()]);
+          "Create actuation residual data.\n\n"
+          ":param model: actuation residual model\n"
+          ":param data: shared data")[bp::with_custodian_and_ward<1, 2, bp::with_custodian_and_ward<1, 3>>()]);
 }
 
 }  // namespace python
