+++ bindings/python/crocoddyl/multibody/multibody.cpp
@@ -1,7 +1,7 @@
 ///////////////////////////////////////////////////////////////////////////////
 // BSD 3-Clause License
 //
-// Copyright (C) 2019-2022, University of Edinburgh
+// Copyright (C) 2019-2022, University of Edinburgh, Heriot-Watt University
 // Copyright note valid unless otherwise stated in individual files.
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
@@ -29,11 +29,11 @@ void exposeMultibody() {
   exposeDataCollectorContacts();
   exposeDataCollectorImpulses();
   exposeDifferentialActionFreeFwdDynamics();
-  exposeDifferentialActionFreeInvDynamicsCondensed();
   exposeDifferentialActionFreeInvDynamics();
+  exposeDifferentialActionFreeInvDynamicsRedundant();
   exposeDifferentialActionContactFwdDynamics();
   exposeDifferentialActionContactInvDynamics();
-  exposeDifferentialActionContactInvDynamicsCondensed();
+  exposeDifferentialActionContactInvDynamicsRedundant();
   exposeActionImpulseFwdDynamics();
   exposeResidualState();
   exposeResidualCentroidalMomentum();
