+++ bindings/python/crocoddyl/multibody/multibody.hpp
@@ -2,6 +2,7 @@
 // BSD 3-Clause License
 //
 // Copyright (C) 2019-2022, LAAS-CNRS, University of Edinburgh
+//                          Heriot-Watt University
 // Copyright note valid unless otherwise stated in individual files.
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
@@ -32,11 +33,11 @@ void exposeDataCollectorMultibody();
 void exposeDataCollectorContacts();
 void exposeDataCollectorImpulses();
 void exposeDifferentialActionFreeFwdDynamics();
-void exposeDifferentialActionFreeInvDynamicsCondensed();
 void exposeDifferentialActionFreeInvDynamics();
+void exposeDifferentialActionFreeInvDynamicsRedundant();
 void exposeDifferentialActionContactFwdDynamics();
 void exposeDifferentialActionContactInvDynamics();
-void exposeDifferentialActionContactInvDynamicsCondensed();
+void exposeDifferentialActionContactInvDynamicsRedundant();
 void exposeActionImpulseFwdDynamics();
 void exposeResidualState();
 void exposeResidualCentroidalMomentum();
