+++ include/crocoddyl/multibody/actions/contact-invdyn.hpp
@@ -1,8 +1,7 @@
 ///////////////////////////////////////////////////////////////////////////////
 // BSD 3-Clause License
 //
-// Copyright (C) 2021-2022, Heriot-Watt University, University of Edinburgh,
-//                          University of Pisa
+// Copyright (C) 2021-2022, Heriot-Watt University, University of Edinburgh
 // Copyright note valid unless otherwise stated in individual files.
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
@@ -24,11 +23,11 @@ namespace crocoddyl {
 /**
  * @brief Differential action model for contact inverse dynamics in multibody systems.
  *
- * This class implements forward kinematic with an inverse-dynamics equality constraint computed using the Recursive
- * Newton Euler Algorithm (RNEA) and contact holonomic constraints (defined at the acceleration level). The stack of
- * cost and constraint functions are implemented in `CostModelSumTpl` and `ConstraintModelManagerTpl`, respectively.
- * The acceleration, torques and contact forces are decision variables defined as the control inputs, and the RNEA
- * and contact constraint are under the name `rnea` and its frame name, thus the user is not allow to use it.
+ * This class implements forward kinematic with contact holonomic constraints (defined at the acceleration level) and
+ * inverse-dynamics computation using the Recursive Newton Euler Algorithm (RNEA). The stack of cost and constraint
+ * functions are implemented in `CostModelSumTpl` and `ConstraintModelManagerTpl`, respectively.
+ * The acceleration and contact forces are decision variables defined as the control inputs, and the under-actuation
+ * and contact constraint are under the name `tau` and its frame name, thus the user is not allow to use it.
  *
  * Additionally, it is important remark that `calcDiff()` computes the derivatives using the latest stored values by
  * `calc()`. Thus, we need to run first `calc()`.
@@ -59,11 +58,7 @@ class DifferentialActionModelContactInvDynamicsTpl : public DifferentialActionMo
    * @brief Initialize the contact inverse-dynamics action model
    *
    * It describes the kinematic evolution of the multibody system with contacts,
-   * and imposes inverse-dynamics and contact-accelerations (equality) constraints.
-   * Additionally, it computes the cost and extra constraint values associated to
-   * this state and control pair. Note that the name `rnea` in the `ConstraintModelManagerTpl`
-   * is reserved to store the inverse-dynamics constraint. We also reserve the name of the
-   * contact frame to define the contact-acceleration constraints.
+   * and computes the needed torques using inverse-dynamics.
    *
    * @param[in] state      State of the multibody system
    * @param[in] actuation  Actuation model
@@ -193,21 +188,18 @@ class DifferentialActionModelContactInvDynamicsTpl : public DifferentialActionMo
 
  public:
   /**
-   * @brief RNEA residual
+   * @brief Actuation residual
    *
-   * This residual function is defined as \f$\mathbf{r} = \boldsymbol{\tau} -
-   * \mathrm{RNEA}(\mathbf{q},\mathbf{v},\dot{\mathbf{a}},\boldsymbol{\lambda})\f$, where \f$\boldsymbol{\tau}\f$
-   * is extracted from the control vector and \f$\mathrm{RNEA}\f$ evaluates the joint torque using \f$\mathbf{q},
-   * \mathbf{v}, \dot{\mathbf{a}}, \boldsymbol{\lambda}\f$ values. Furthermore, the Jacobians of the residual
-   * function are computed analytically. This is used by `ConstraintModelManagerTpl` inside parent
-   * `DifferentialActionModelContactInvDynamicsTpl` class.
+   * This residual function enforces the torques of under-actuated joints (e.g., floating-base joints) to be zero.
+   * We compute these torques and their derivatives using RNEA inside
+   * `DifferentialActionModelContactInvDynamicsTpl`.
    *
    * As described in `ResidualModelAbstractTpl`, the residual value and its Jacobians are calculated by `calc` and
    * `calcDiff`, respectively.
    *
    * \sa `ResidualModelAbstractTpl`, `calc()`, `calcDiff()`, `createData()`
    */
-  class ResidualModelRnea : public ResidualModelAbstractTpl<_Scalar> {
+  class ResidualModelActuation : public ResidualModelAbstractTpl<_Scalar> {
    public:
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
@@ -220,66 +212,80 @@ class DifferentialActionModelContactInvDynamicsTpl : public DifferentialActionMo
     typedef typename MathBase::VectorXs VectorXs;
 
     /**
-     * @brief Initialize the RNEA residual model
+     * @brief Initialize the actuation residual model
      *
      * @param[in] state  State of the multibody system
+     * @param[in] nu     Dimension of the joint torques
      * @param[in] nc     Dimension of all the contacts
-     * @param[in] nu     Dimension of the control vector
      */
-    ResidualModelRnea(boost::shared_ptr<StateMultibody> state, const std::size_t nc, const std::size_t nu)
-        : Base(state, state->get_nv(), state->get_nv() + nu + nc, true, true, true), nc_(nc), na_(nu) {}
-    virtual ~ResidualModelRnea() {}
+    ResidualModelActuation(boost::shared_ptr<StateMultibody> state, const std::size_t nu, const std::size_t nc)
+        : Base(state, state->get_nv() - nu, state->get_nv() + nc, true, true, true), na_(nu), nc_(nc) {}
+    virtual ~ResidualModelActuation() {}
 
     /**
-     * @brief Compute the RNEA residual
+     * @brief Compute the actuation residual
      *
-     * @param[in] data  RNEA residual data
+     * @param[in] data  Actuation residual data
      * @param[in] x     State point \f$\mathbf{x}\in\mathbb{R}^{ndx}\f$
      * @param[in] u     Control input \f$\mathbf{u}\in\mathbb{R}^{nv+nu}\f$
      */
     virtual void calc(const boost::shared_ptr<ResidualDataAbstract>& data, const Eigen::Ref<const VectorXs>&,
                       const Eigen::Ref<const VectorXs>&) {
-      typename Data::ResidualDataRnea* d = static_cast<typename Data::ResidualDataRnea*>(data.get());
-      data->r = d->pinocchio->tau - d->actuation->tau;
+      typename Data::ResidualDataActuation* d = static_cast<typename Data::ResidualDataActuation*>(data.get());
+      // Update the under-actuation set and residual
+      std::size_t nrow = 0;
+      for (std::size_t k = 0; k < static_cast<std::size_t>(d->actuation->tau_set.size()); ++k) {
+        if (!d->actuation->tau_set[k]) {
+          data->r(nrow) = d->pinocchio->tau(k);
+          nrow += 1;
+        }
+      }
     }
 
     /**
-     * @brief Compute the derivatives of the RNEA residual
+     * @brief Compute the derivatives of the actuation residual
      *
-     * @param[in] data  RNEA residual data
+     * @param[in] data  Actuation residual data
      * @param[in] x     State point \f$\mathbf{x}\in\mathbb{R}^{ndx}\f$
      * @param[in] u     Control input \f$\mathbf{u}\in\mathbb{R}^{nu}\f$
      */
     virtual void calcDiff(const boost::shared_ptr<ResidualDataAbstract>& data, const Eigen::Ref<const VectorXs>&,
                           const Eigen::Ref<const VectorXs>&) {
-      typename Data::ResidualDataRnea* d = static_cast<typename Data::ResidualDataRnea*>(data.get());
+      typename Data::ResidualDataActuation* d = static_cast<typename Data::ResidualDataActuation*>(data.get());
+      std::size_t nrow = 0;
       const std::size_t nv = state_->get_nv();
-      data->Rx.leftCols(nv) = d->pinocchio->dtau_dq;
-      data->Rx.rightCols(nv) = d->pinocchio->dtau_dv;
-      data->Rx -= d->actuation->dtau_dx;
-      data->Ru.leftCols(nv) = d->pinocchio->M;
-      data->Ru.middleCols(nv, na_) = -d->actuation->dtau_du;
-      data->Ru.rightCols(nc_) = -d->contact->Jc.transpose();
+      d->dtau_dx.leftCols(nv) = d->pinocchio->dtau_dq;
+      d->dtau_dx.rightCols(nv) = d->pinocchio->dtau_dv;
+      d->dtau_dx -= d->actuation->dtau_dx;
+      d->dtau_du.leftCols(nv) = d->pinocchio->M;
+      d->dtau_du.rightCols(nc_) = -d->contact->Jc.transpose();
+      for (std::size_t k = 0; k < static_cast<std::size_t>(d->actuation->tau_set.size()); ++k) {
+        if (!d->actuation->tau_set[k]) {
+          d->Rx.row(nrow) = d->dtau_dx.row(k);
+          d->Ru.row(nrow) = d->dtau_du.row(k);
+          nrow += 1;
+        }
+      }
     }
 
     /**
-     * @brief Create the RNEA residual data
+     * @brief Create the actuation residual data
      *
-     * @return RNEA residual data
+     * @return Actuation residual data
      */
     virtual boost::shared_ptr<ResidualDataAbstract> createData(DataCollectorAbstract* const data) {
-      return boost::allocate_shared<typename Data::ResidualDataRnea>(
-          Eigen::aligned_allocator<typename Data::ResidualDataRnea>(), this, data);
+      return boost::allocate_shared<typename Data::ResidualDataActuation>(
+          Eigen::aligned_allocator<typename Data::ResidualDataActuation>(), this, data);
     }
 
     /**
-     * @brief Print relevant information of the RNEA residual model
+     * @brief Print relevant information of the actuation residual model
      *
      * @param[out] os  Output stream object
      */
     virtual void print(std::ostream& os) const {
-      os << "ResidualModelRnea {nx=" << state_->get_nx() << ", ndx=" << state_->get_ndx() << ", nu=" << nu_
-         << "nc=" << nc_;
+      os << "ResidualModelActuation {nx=" << state_->get_nx() << ", ndx=" << state_->get_ndx() << ", nu=" << nu_
+         << ", na=" << na_ << "}";
     }
 
    protected:
@@ -287,8 +293,8 @@ class DifferentialActionModelContactInvDynamicsTpl : public DifferentialActionMo
     using Base::state_;
 
    private:
-    std::size_t nc_;  //!< Number of all contacts
     std::size_t na_;  //!< Number of actuated joints
+    std::size_t nc_;  //!< Number of contacts
   };
 
  public:
@@ -325,11 +331,10 @@ class DifferentialActionModelContactInvDynamicsTpl : public DifferentialActionMo
      * @param[in] id     Contact frame id
      * @param[in] nr     Dimension of the contact-acceleration residual
      * @param[in] nc     Dimension of all contacts
-     * @param[in] nu     Dimension of the control vector
      */
     ResidualModelContact(boost::shared_ptr<StateMultibody> state, const pinocchio::FrameIndex id, const std::size_t nr,
-                         const std::size_t nc, const std::size_t nu)
-        : Base(state, nr, state->get_nv() + nu + nc, true, true, true), id_(id) {}
+                         const std::size_t nc)
+        : Base(state, nr, state->get_nv() + nc, true, true, true), id_(id) {}
     virtual ~ResidualModelContact() {}
 
     /**
@@ -385,10 +390,11 @@ struct DifferentialActionDataContactInvDynamicsTpl : public DifferentialActionDa
   typedef _Scalar Scalar;
   typedef MathBaseTpl<Scalar> MathBase;
   typedef DifferentialActionDataAbstractTpl<Scalar> Base;
-  typedef ContactModelMultipleTpl<Scalar> ContactModelMultiple;
-  typedef DataCollectorActMultibodyInContactTpl<Scalar> DataCollectorActMultibodyInContact;
+  typedef JointDataAbstractTpl<Scalar> JointDataAbstract;
+  typedef DataCollectorJointActMultibodyInContactTpl<Scalar> DataCollectorJointActMultibodyInContact;
   typedef CostDataSumTpl<Scalar> CostDataSum;
   typedef ConstraintDataManagerTpl<Scalar> ConstraintDataManager;
+  typedef ContactModelMultipleTpl<Scalar> ContactModelMultiple;
   typedef typename MathBase::VectorXs VectorXs;
   typedef typename MathBase::MatrixXs MatrixXs;
 
@@ -396,15 +402,17 @@ struct DifferentialActionDataContactInvDynamicsTpl : public DifferentialActionDa
   explicit DifferentialActionDataContactInvDynamicsTpl(Model<Scalar>* const model)
       : Base(model),
         pinocchio(pinocchio::DataTpl<Scalar>(model->get_pinocchio())),
-        multibody(&pinocchio, model->get_actuation()->createData(), model->get_contacts()->createData(&pinocchio)),
+        multibody(&pinocchio, model->get_actuation()->createData(),
+                  boost::make_shared<JointDataAbstract>(model->get_state(), model->get_actuation(), model->get_nu()),
+                  model->get_contacts()->createData(&pinocchio)),
         tmp_xstatic(model->get_state()->get_nx()),
         tmp_rstatic(model->get_actuation()->get_nu() + model->get_contacts()->get_nc()),
         tmp_Jstatic(model->get_state()->get_nv(), model->get_actuation()->get_nu() + model->get_contacts()->get_nc()) {
     // Set constant values for Fu, df_dx, and df_du
-    const std::size_t na = model->get_actuation()->get_nu();
     const std::size_t nv = model->get_state()->get_nv();
     const std::size_t nc = model->get_contacts()->get_nc_total();
     Fu.leftCols(nv).diagonal().setOnes();
+    multibody.joint->da_du.leftCols(nv).diagonal().setOnes();
     MatrixXs df_dx = MatrixXs::Zero(nc, model->get_state()->get_ndx());
     MatrixXs df_du = MatrixXs::Zero(nc, model->get_nu());
     std::size_t fid = 0;
@@ -412,11 +420,10 @@ struct DifferentialActionDataContactInvDynamicsTpl : public DifferentialActionDa
              model->get_contacts()->get_contacts().begin();
          it != model->get_contacts()->get_contacts().end(); ++it) {
       const std::size_t nc = it->second->contact->get_nc();
-      df_du.block(fid, nv + na + fid, nc, nc).diagonal().setOnes();
+      df_du.block(fid, nv + fid, nc, nc).diagonal().setOnes();
       fid += nc;
     }
     model->get_contacts()->updateForceDiff(multibody.contacts, df_dx, df_du);
-
     costs = model->get_costs()->createData(&multibody);
     constraints = model->get_constraints()->createData(&multibody);
     costs->shareMemory(this);
@@ -428,7 +435,7 @@ struct DifferentialActionDataContactInvDynamicsTpl : public DifferentialActionDa
   }
 
   pinocchio::DataTpl<Scalar> pinocchio;                  //!< Pinocchio data
-  DataCollectorActMultibodyInContact multibody;          //!< Multibody data
+  DataCollectorJointActMultibodyInContact multibody;     //!< Multibody data
   boost::shared_ptr<CostDataSum> costs;                  //!< Costs data
   boost::shared_ptr<ConstraintDataManager> constraints;  //!< Constraints data
   VectorXs tmp_xstatic;                                  //!< State point used for computing the quasi-static input
@@ -446,7 +453,7 @@ struct DifferentialActionDataContactInvDynamicsTpl : public DifferentialActionDa
   using Base::r;
   using Base::xout;
 
-  struct ResidualDataRnea : public ResidualDataAbstractTpl<_Scalar> {
+  struct ResidualDataActuation : public ResidualDataAbstractTpl<_Scalar> {
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
     typedef _Scalar Scalar;
@@ -454,9 +461,15 @@ struct DifferentialActionDataContactInvDynamicsTpl : public DifferentialActionDa
     typedef ResidualDataAbstractTpl<Scalar> Base;
     typedef DataCollectorAbstractTpl<Scalar> DataCollectorAbstract;
     typedef DataCollectorActMultibodyInContactTpl<Scalar> DataCollectorActMultibodyInContact;
+    typedef ActuationDataAbstractTpl<Scalar> ActuationDataAbstract;
+    typedef ContactDataMultipleTpl<Scalar> ContactDataMultiple;
+    typedef typename MathBase::MatrixXs MatrixXs;
 
     template <template <typename Scalar> class Model>
-    ResidualDataRnea(Model<Scalar>* const model, DataCollectorAbstract* const data) : Base(model, data) {
+    ResidualDataActuation(Model<Scalar>* const model, DataCollectorAbstract* const data)
+        : Base(model, data),
+          dtau_dx(model->get_state()->get_nv(), model->get_state()->get_ndx()),
+          dtau_du(model->get_state()->get_nv(), model->get_nu()) {
       // Check that proper shared data has been passed
       DataCollectorActMultibodyInContact* d = dynamic_cast<DataCollectorActMultibodyInContact*>(shared);
       if (d == NULL) {
@@ -466,11 +479,15 @@ struct DifferentialActionDataContactInvDynamicsTpl : public DifferentialActionDa
       pinocchio = d->pinocchio;
       actuation = d->actuation.get();
       contact = d->contacts.get();
+      dtau_dx.setZero();
+      dtau_du.setZero();
     }
 
-    pinocchio::DataTpl<Scalar>* pinocchio;        //!< Pinocchio data
-    ActuationDataAbstractTpl<Scalar>* actuation;  //!< Actuation data
-    ContactDataMultipleTpl<Scalar>* contact;      //!< Contact data
+    pinocchio::DataTpl<Scalar>* pinocchio;  //!< Pinocchio data
+    ActuationDataAbstract* actuation;       //!< Actuation data
+    ContactDataMultiple* contact;           //!< Contact data
+    MatrixXs dtau_dx;
+    MatrixXs dtau_du;
     using Base::r;
     using Base::Ru;
     using Base::Rx;
