+++ include/crocoddyl/multibody/actions/free-invdyn.hpp
@@ -1,8 +1,7 @@
 ///////////////////////////////////////////////////////////////////////////////
 // BSD 3-Clause License
 //
-// Copyright (C) 2021-2022, Heriot-Watt University, University of Edinburgh,
-//                          University of Pisa
+// Copyright (C) 2021-2022, Heriot-Watt University, University of Edinburgh
 // Copyright note valid unless otherwise stated in individual files.
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
@@ -27,11 +26,12 @@ namespace crocoddyl {
 /**
  * @brief Differential action model for free inverse dynamics in multibody systems.
  *
- * This class implements forward kinematic with an inverse-dynamics equality constraint computed using the Recursive
- * Newton Euler Algorithm (RNEA). The stack of cost and constraint functions are implemented in
- * `CostModelSumTpl` and `ConstraintModelManagerTpl`, respectively. The acceleration and the torques are decision
- * variables defined as the control inputs, and the RNEA constraint is under the name `rnea`, thus the user is not
- * allow to use it.
+ * This class implements forward kinematic with an inverse-dynamics computed using the Recursive Newton Euler
+ * Algorithm (RNEA). The stack of cost and constraint functions are implemented in
+ * `CostModelSumTpl` and `ConstraintModelManagerTpl`, respectively. The accelerations are the decision
+ * variables defined as the control inputs, and the under-actuation constraint is under the name `tau`, thus the
+ *  user is not allow to use it.
+ *
  *
  * \sa `DifferentialActionModelAbstractTpl`, `calc()`, `calcDiff()`, `createData()`
  */
@@ -55,11 +55,8 @@ class DifferentialActionModelFreeInvDynamicsTpl : public DifferentialActionModel
   /**
    * @brief Initialize the free inverse-dynamics action model
    *
-   * It describes the kinematic evolution of the multibody system without any contact,
-   * and imposes an inverse-dynamics (equality) constraint. Additionally, it computes
-   * the cost and extra constraint values associated to this state and control pair.
-   * Note that the name `rnea` in the `ConstraintModelManagerTpl` is reserved to store
-   * the inverse-dynamics constraint.
+   * It describes the kinematic evolution of the multibody system and computes the needed torques
+   * using inverse dynamics.
    *
    * @param[in] state      State of the multibody system
    * @param[in] actuation  Actuation model
@@ -179,20 +176,18 @@ class DifferentialActionModelFreeInvDynamicsTpl : public DifferentialActionModel
 
  public:
   /**
-   * @brief RNEA residual
+   * @brief Actuation residual
    *
-   * This residual function is defined as \f$\mathbf{r} = \boldsymbol{\tau} -
-   * \mathrm{RNEA}(\mathbf{q},\mathbf{v},\dot{\mathbf{a}})\f$, where \f$\boldsymbol{\tau}\f$ is extracted from the
-   * control vector and \f$\mathrm{RNEA}\f$ evaluates the joint torque using \f$\mathbf{q}, \mathbf{v},
-   * \dot{\mathbf{a}}\f$ values. Furthermore, the Jacobians of the residual function are computed analytically.
-   * This is used by `ConstraintModelManagerTpl` inside parent `DifferentialActionModelFreeInvDynamicsTpl` class.
+   * This residual function enforces the torques of under-actuated joints (e.g., floating-base joints) to be zero.
+   * We compute these torques and their derivatives using RNEA inside
+   * `DifferentialActionModelFreeInvDynamicsTpl`.
    *
    * As described in `ResidualModelAbstractTpl`, the residual value and its Jacobians are calculated by `calc` and
    * `calcDiff`, respectively.
    *
    * \sa `ResidualModelAbstractTpl`, `calc()`, `calcDiff()`, `createData()`
    */
-  class ResidualModelRnea : public ResidualModelAbstractTpl<_Scalar> {
+  class ResidualModelActuation : public ResidualModelAbstractTpl<_Scalar> {
    public:
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
@@ -202,71 +197,86 @@ class DifferentialActionModelFreeInvDynamicsTpl : public DifferentialActionModel
     typedef StateMultibodyTpl<Scalar> StateMultibody;
     typedef ResidualDataAbstractTpl<Scalar> ResidualDataAbstract;
     typedef DataCollectorAbstractTpl<Scalar> DataCollectorAbstract;
+    typedef ActuationModelAbstractTpl<Scalar> ActuationModelAbstract;
     typedef typename MathBase::VectorXs VectorXs;
+    typedef typename MathBase::MatrixXs MatrixXs;
 
     /**
-     * @brief Initialize the RNEA residual model
+     * @brief Initialize the actuation residual model
      *
      * @param[in] state  State of the multibody system
-     * @param[in] nu     Dimension of the control vector
+     * @param[in] nu     Dimension of the joint torques
      */
-    ResidualModelRnea(boost::shared_ptr<StateMultibody> state, const std::size_t nu)
-        : Base(state, state->get_nv(), state->get_nv() + nu, true, true, true), na_(nu) {}
-    virtual ~ResidualModelRnea() {}
+    ResidualModelActuation(boost::shared_ptr<StateMultibody> state, const std::size_t nu)
+        : Base(state, state->get_nv() - nu, state->get_nv(), true, true, true), na_(nu) {}
+    virtual ~ResidualModelActuation() {}
 
     /**
-     * @brief Compute the RNEA residual
+     * @brief Compute the actuation residual
      *
-     * @param[in] data  RNEA residual data
+     * @param[in] data  Actuation residual data
      * @param[in] x     State point \f$\mathbf{x}\in\mathbb{R}^{ndx}\f$
      * @param[in] u     Control input \f$\mathbf{u}\in\mathbb{R}^{nv+nu}\f$
      */
     virtual void calc(const boost::shared_ptr<ResidualDataAbstract>& data, const Eigen::Ref<const VectorXs>&,
                       const Eigen::Ref<const VectorXs>&) {
-      typename Data::ResidualDataRnea* d = static_cast<typename Data::ResidualDataRnea*>(data.get());
-      data->r = d->pinocchio->tau - d->actuation->tau;
+      typename Data::ResidualDataActuation* d = static_cast<typename Data::ResidualDataActuation*>(data.get());
+      // Update the under-actuation set and residual
+      std::size_t nrow = 0;
+      for (std::size_t k = 0; k < static_cast<std::size_t>(d->actuation->tau_set.size()); ++k) {
+        if (!d->actuation->tau_set[k]) {
+          data->r(nrow) = d->pinocchio->tau(k);
+          nrow += 1;
+        }
+      }
     }
 
     /**
-     * @brief Compute the derivatives of the RNEA residual
+     * @brief Compute the derivatives of the actuation residual
      *
-     * @param[in] data  RNEA residual data
+     * @param[in] data  Actuation residual data
      * @param[in] x     State point \f$\mathbf{x}\in\mathbb{R}^{ndx}\f$
      * @param[in] u     Control input \f$\mathbf{u}\in\mathbb{R}^{nu}\f$
      */
     virtual void calcDiff(const boost::shared_ptr<ResidualDataAbstract>& data, const Eigen::Ref<const VectorXs>&,
                           const Eigen::Ref<const VectorXs>&) {
-      typename Data::ResidualDataRnea* d = static_cast<typename Data::ResidualDataRnea*>(data.get());
+      typename Data::ResidualDataActuation* d = static_cast<typename Data::ResidualDataActuation*>(data.get());
+      std::size_t nrow = 0;
       const std::size_t nv = state_->get_nv();
-      data->Rx.leftCols(nv) = d->pinocchio->dtau_dq;
-      data->Rx.rightCols(nv) = d->pinocchio->dtau_dv;
-      data->Rx -= d->actuation->dtau_dx;
-      data->Ru.leftCols(nv) = d->pinocchio->M;
-      data->Ru.rightCols(na_) = -d->actuation->dtau_du;
+      d->dtau_dx.leftCols(nv) = d->pinocchio->dtau_dq;
+      d->dtau_dx.rightCols(nv) = d->pinocchio->dtau_dv;
+      d->dtau_dx -= d->actuation->dtau_dx;
+      for (std::size_t k = 0; k < static_cast<std::size_t>(d->actuation->tau_set.size()); ++k) {
+        if (!d->actuation->tau_set[k]) {
+          d->Rx.row(nrow) = d->dtau_dx.row(k);
+          d->Ru.row(nrow) = d->pinocchio->M.row(k);
+          nrow += 1;
+        }
+      }
     }
 
     /**
-     * @brief Create the RNEA residual data
+     * @brief Create the actuation residual data
      *
-     * @return RNEA residual data
+     * @return Actuation residual data
      */
     virtual boost::shared_ptr<ResidualDataAbstract> createData(DataCollectorAbstract* const data) {
-      return boost::allocate_shared<typename Data::ResidualDataRnea>(
-          Eigen::aligned_allocator<typename Data::ResidualDataRnea>(), this, data);
+      return boost::allocate_shared<typename Data::ResidualDataActuation>(
+          Eigen::aligned_allocator<typename Data::ResidualDataActuation>(), this, data);
     }
 
     /**
-     * @brief Print relevant information of the RNEA residual model
+     * @brief Print relevant information of the actuation residual model
      *
      * @param[out] os  Output stream object
      */
     virtual void print(std::ostream& os) const {
-      os << "ResidualModelRnea {nx=" << state_->get_nx() << ", ndx=" << state_->get_ndx() << ", nu=" << nu_
+      os << "ResidualModelActuation {nx=" << state_->get_nx() << ", ndx=" << state_->get_ndx() << ", nu=" << nu_
          << ", na=" << na_ << "}";
     }
 
    protected:
-    std::size_t na_;  //!< Number of actuated joints
+    std::size_t na_;  //!< Dimension of the joint torques
     using Base::nu_;
     using Base::state_;
   };
@@ -278,27 +288,34 @@ struct DifferentialActionDataFreeInvDynamicsTpl : public DifferentialActionDataA
   typedef _Scalar Scalar;
   typedef MathBaseTpl<Scalar> MathBase;
   typedef DifferentialActionDataAbstractTpl<Scalar> Base;
+  typedef JointDataAbstractTpl<Scalar> JointDataAbstract;
+  typedef DataCollectorJointActMultibodyTpl<Scalar> DataCollectorJointActMultibody;
+  typedef CostDataSumTpl<Scalar> CostDataSum;
+  typedef ConstraintDataManagerTpl<Scalar> ConstraintDataManager;
   typedef typename MathBase::VectorXs VectorXs;
 
   template <template <typename Scalar> class Model>
   explicit DifferentialActionDataFreeInvDynamicsTpl(Model<Scalar>* const model)
       : Base(model),
         pinocchio(pinocchio::DataTpl<Scalar>(model->get_pinocchio())),
-        multibody(&pinocchio, model->get_actuation()->createData()),
+        multibody(&pinocchio, model->get_actuation()->createData(),
+                  boost::make_shared<JointDataAbstract>(model->get_state(), model->get_actuation(), model->get_nu())),
         costs(model->get_costs()->createData(&multibody)),
         constraints(model->get_constraints()->createData(&multibody)),
         tmp_xstatic(model->get_state()->get_nx()) {
+    const std::size_t nv = model->get_state()->get_nv();
+    Fu.leftCols(nv).diagonal().setOnes();
+    multibody.joint->da_du.leftCols(nv).diagonal().setOnes();
     costs->shareMemory(this);
     constraints->shareMemory(this);
-    Fu.leftCols(model->get_state()->get_nv()).diagonal().array() = 1;
     tmp_xstatic.setZero();
   }
 
-  pinocchio::DataTpl<Scalar> pinocchio;                              //!< Pinocchio data
-  DataCollectorActMultibodyTpl<Scalar> multibody;                    //!< Multibody data
-  boost::shared_ptr<CostDataSumTpl<Scalar> > costs;                  //!< Costs data
-  boost::shared_ptr<ConstraintDataManagerTpl<Scalar> > constraints;  //!< Constraints data
-  VectorXs tmp_xstatic;  //!< quasistatic state point (velocity has to be zero)
+  pinocchio::DataTpl<Scalar> pinocchio;                  //!< Pinocchio data
+  DataCollectorJointActMultibody multibody;              //!< Multibody data
+  boost::shared_ptr<CostDataSum> costs;                  //!< Costs data
+  boost::shared_ptr<ConstraintDataManager> constraints;  //!< Constraints data
+  VectorXs tmp_xstatic;                                  //!< State point used for computing the quasi-static input
   using Base::cost;
   using Base::Fu;
   using Base::Fx;
@@ -310,31 +327,35 @@ struct DifferentialActionDataFreeInvDynamicsTpl : public DifferentialActionDataA
   using Base::r;
   using Base::xout;
 
-  struct ResidualDataRnea : public ResidualDataAbstractTpl<_Scalar> {
+  struct ResidualDataActuation : public ResidualDataAbstractTpl<_Scalar> {
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
     typedef _Scalar Scalar;
     typedef MathBaseTpl<Scalar> MathBase;
     typedef ResidualDataAbstractTpl<Scalar> Base;
     typedef DataCollectorAbstractTpl<Scalar> DataCollectorAbstract;
+    typedef DataCollectorActMultibodyTpl<Scalar> DataCollectorActMultibody;
+    typedef ActuationDataAbstractTpl<Scalar> ActuationDataAbstract;
+    typedef typename MathBase::MatrixXs MatrixXs;
 
     template <template <typename Scalar> class Model>
-    ResidualDataRnea(Model<Scalar>* const model, DataCollectorAbstract* const data) : Base(model, data) {
+    ResidualDataActuation(Model<Scalar>* const model, DataCollectorAbstract* const data)
+        : Base(model, data), dtau_dx(model->get_state()->get_nv(), model->get_state()->get_ndx()) {
+      dtau_dx.setZero();
       // Check that proper shared data has been passed
-      DataCollectorActMultibodyTpl<Scalar>* d = dynamic_cast<DataCollectorActMultibodyTpl<Scalar>*>(shared);
+      DataCollectorActMultibody* d = dynamic_cast<DataCollectorActMultibody*>(shared);
       if (d == NULL) {
         throw_pretty("Invalid argument: the shared data should be derived from DataCollectorActMultibody");
       }
-      const std::size_t na = Ru.cols() - model->get_state()->get_nv();
-      Ru.rightCols(na) = -d->actuation->dtau_du;
 
       // Avoids data casting at runtime
       pinocchio = d->pinocchio;
       actuation = d->actuation.get();
     }
 
-    pinocchio::DataTpl<Scalar>* pinocchio;        //!< Pinocchio data
-    ActuationDataAbstractTpl<Scalar>* actuation;  //!< Actuation data
+    pinocchio::DataTpl<Scalar>* pinocchio;  //!< Pinocchio data
+    ActuationDataAbstract* actuation;       //!< Actuation data
+    MatrixXs dtau_dx;
     using Base::r;
     using Base::Ru;
     using Base::Rx;
