+++ unittest/factory/diff_action.hpp
@@ -1,7 +1,8 @@
 ///////////////////////////////////////////////////////////////////////////////
 // BSD 3-Clause License
 //
-// Copyright (C) 2019-2021, University of Edinburgh, CTU, INRIA
+// Copyright (C) 2019-2022, University of Edinburgh, CTU, INRIA,
+//                          Heriot-Watt University
 // Copyright note valid unless otherwise stated in individual files.
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
@@ -18,10 +19,10 @@
 #include "crocoddyl/core/numdiff/diff-action.hpp"
 #include "crocoddyl/multibody/actions/free-fwddyn.hpp"
 #include "crocoddyl/multibody/actions/free-invdyn.hpp"
-#include "crocoddyl/multibody/actions/free-invdyn-condensed.hpp"
+#include "crocoddyl/multibody/actions/free-invdyn-redundant.hpp"
 #include "crocoddyl/multibody/actions/contact-fwddyn.hpp"
-#include "crocoddyl/multibody/actions/contact-invdyn-condensed.hpp"
 #include "crocoddyl/multibody/actions/contact-invdyn.hpp"
+#include "crocoddyl/multibody/actions/contact-invdyn-redundant.hpp"
 
 namespace crocoddyl {
 namespace unittest {
@@ -33,12 +34,12 @@ struct DifferentialActionModelTypes {
     DifferentialActionModelFreeFwdDynamics_Hector,
     DifferentialActionModelFreeFwdDynamics_TalosArm,
     DifferentialActionModelFreeFwdDynamics_TalosArm_Squashed,
-    DifferentialActionModelFreeInvDynamicsCondensed_Hector,
-    DifferentialActionModelFreeInvDynamicsCondensed_TalosArm,
-    DifferentialActionModelFreeInvDynamicsCondensed_TalosArm_Squashed,
     DifferentialActionModelFreeInvDynamics_Hector,
     DifferentialActionModelFreeInvDynamics_TalosArm,
     DifferentialActionModelFreeInvDynamics_TalosArm_Squashed,
+    DifferentialActionModelFreeInvDynamicsRedundant_Hector,
+    DifferentialActionModelFreeInvDynamicsRedundant_TalosArm,
+    DifferentialActionModelFreeInvDynamicsRedundant_TalosArm_Squashed,
     DifferentialActionModelContactFwdDynamics_TalosArm,
     DifferentialActionModelContact2DFwdDynamics_TalosArm,
     DifferentialActionModelContactFwdDynamics_HyQ,
@@ -47,18 +48,18 @@ struct DifferentialActionModelTypes {
     DifferentialActionModelContact2DFwdDynamicsWithFriction_TalosArm,
     DifferentialActionModelContactFwdDynamicsWithFriction_HyQ,
     DifferentialActionModelContactFwdDynamicsWithFriction_Talos,
-    DifferentialActionModelContactInvDynamicsCondensed_TalosArm,
-    DifferentialActionModelContactInvDynamicsCondensed_HyQ,
-    DifferentialActionModelContactInvDynamicsCondensed_Talos,
-    DifferentialActionModelContactInvDynamicsCondensedWithFriction_TalosArm,
-    DifferentialActionModelContactInvDynamicsCondensedWithFriction_HyQ,
-    DifferentialActionModelContactInvDynamicsCondensedWithFriction_Talos,
     DifferentialActionModelContactInvDynamics_TalosArm,
     DifferentialActionModelContactInvDynamics_HyQ,
     DifferentialActionModelContactInvDynamics_Talos,
     DifferentialActionModelContactInvDynamicsWithFriction_TalosArm,
     DifferentialActionModelContactInvDynamicsWithFriction_HyQ,
     DifferentialActionModelContactInvDynamicsWithFriction_Talos,
+    DifferentialActionModelContactInvDynamicsRedundant_TalosArm,
+    DifferentialActionModelContactInvDynamicsRedundant_HyQ,
+    DifferentialActionModelContactInvDynamicsRedundant_Talos,
+    DifferentialActionModelContactInvDynamicsRedundantWithFriction_TalosArm,
+    DifferentialActionModelContactInvDynamicsRedundantWithFriction_HyQ,
+    DifferentialActionModelContactInvDynamicsRedundantWithFriction_Talos,
     NbDifferentialActionModelTypes
   };
   static std::vector<Type> init_all() {
@@ -86,19 +87,19 @@ class DifferentialActionModelFactory {
   boost::shared_ptr<crocoddyl::DifferentialActionModelFreeFwdDynamics> create_freeFwdDynamics(
       StateModelTypes::Type state_type, ActuationModelTypes::Type actuation_type, bool constraints = true) const;
 
-  boost::shared_ptr<crocoddyl::DifferentialActionModelFreeInvDynamicsCondensed> create_freeInvDynamicsCondensed(
+  boost::shared_ptr<crocoddyl::DifferentialActionModelFreeInvDynamics> create_freeInvDynamics(
       StateModelTypes::Type state_type, ActuationModelTypes::Type actuation_type, bool constraints = true) const;
 
-  boost::shared_ptr<crocoddyl::DifferentialActionModelFreeInvDynamics> create_freeInvDynamics(
+  boost::shared_ptr<crocoddyl::DifferentialActionModelFreeInvDynamicsRedundant> create_freeInvDynamicsRedundant(
       StateModelTypes::Type state_type, ActuationModelTypes::Type actuation_type, bool constraints = true) const;
 
   boost::shared_ptr<crocoddyl::DifferentialActionModelContactFwdDynamics> create_contactFwdDynamics(
       StateModelTypes::Type state_type, ActuationModelTypes::Type actuation_type, bool with_friction = true) const;
 
-  boost::shared_ptr<crocoddyl::DifferentialActionModelContactInvDynamicsCondensed> create_contactInvDynamicsCondensed(
+  boost::shared_ptr<crocoddyl::DifferentialActionModelContactInvDynamics> create_contactInvDynamics(
       StateModelTypes::Type state_type, ActuationModelTypes::Type actuation_type, bool with_friction = true) const;
 
-  boost::shared_ptr<crocoddyl::DifferentialActionModelContactInvDynamics> create_contactInvDynamics(
+  boost::shared_ptr<crocoddyl::DifferentialActionModelContactInvDynamicsRedundant> create_contactInvDynamicsRedundant(
       StateModelTypes::Type state_type, ActuationModelTypes::Type actuation_type, bool with_friction = true) const;
 };
